+++ src/app/services/chat-manager.ts
@@ -6,7 +6,8 @@ import { UserModel } from '../models/user';
 
 // handlers
 import { ChatConversationHandler } from './chat-conversation-handler';
-import { ChatConversationsHandler } from './chat-conversations-handler';
+// import { ChatConversationsHandler } from './chat-conversations-handler';
+import { ConversationsHandlerService } from 'src/app/services/conversations-handler.service';
 import { ChatArchivedConversationsHandler } from './chat-archived-conversations-handler';
 import { ChatContactsSynchronizer } from './chat-contacts-synchronizer';
 import { environment } from '../../environments/environment';
@@ -26,7 +27,7 @@ export class ChatManager {
   private tenant: string;
   public handlers: ChatConversationHandler[];
   private loggedUser: UserModel;
-  public conversationsHandler: ChatConversationsHandler;
+  // public conversationsHandler: ChatConversationsHandler;
   public archivedConversationsHandler: ChatArchivedConversationsHandler;
   public contactsSynchronizer: ChatContactsSynchronizer;
   public openInfoConversation: boolean;
@@ -35,7 +36,8 @@ export class ChatManager {
   constructor(
     public chatContactsSynchronizer: ChatContactsSynchronizer,
     private events: EventsService,
-    public userService: UserService
+    public userService: UserService,
+    public conversationsHandlerService: ConversationsHandlerService
   ) { }
   /**
    * inizializza chatmanager
@@ -124,16 +126,16 @@ export class ChatManager {
    * dispose reference sincronizzazione contatti
    */
   dispose() {
-    console.log(" 1 - setOffAllReferences");
-    this.setOffAllReferences();
-    console.log(" 2 - disposeConversationsHandler");
-    if (this.conversationsHandler) { this.disposeConversationsHandler(); }
-    console.log(" 3 - disposeArchivedConversationsHandler");
+    console.log(' 1 - setOffAllReferences');
+    if(this.handlers) { this.setOffAllReferences(); }
+    console.log(' 2 - disposeConversationsHandler');
+    if (this.conversationsHandlerService) { this.disposeConversationsHandler(); }
+    console.log(' 3 - disposeArchivedConversationsHandler');
     if (this.archivedConversationsHandler) { this.disposeConversationsHandler(); }
-    console.log(" 4 - disposeContactsSynchronizer");
+    console.log(' 4 - disposeContactsSynchronizer');
     if (this.contactsSynchronizer) { this.disposeContactsSynchronizer(); }
-    console.log(" OKK ");
-    this.conversationsHandler = null;
+    console.log(' OKK ');
+    this.conversationsHandlerService = null;
     this.contactsSynchronizer = null;
   }
 
@@ -179,20 +181,18 @@ export class ChatManager {
    */
   goOffLine() {
     this.loggedUser = null;
-    console.log(" 1 - CANCELLO TUTTE LE REFERENCES DI FIREBASE");
+    console.log(' 1 - CANCELLO TUTTE LE REFERENCES DI FIREBASE');
     this.dispose();
-    // console.log(" 4 - PUBBLICO STATO LOGGEDUSER: LOGOUT");
-    // this.events.publish('loggedUser:logout', null);
   }
 
   /// START metodi gestione messaggi conversazione ////
   /**
    * aggiungo la conversazione all'array delle conversazioni
    * chiamato dall'inizialize di dettaglio-conversazione.ts
-   * @param handler 
+   * @param handler
    */
   addConversationHandler(handler) {
-    console.log("CHAT MANAGER -----> addConversationHandler", handler);
+    console.log('CHAT MANAGER -----> addConversationHandler', handler);
     this.handlers.push(handler);
   }
 
@@ -202,7 +202,7 @@ export class ChatManager {
    * @param conversationId 
    */
   removeConversationHandler(conversationId) {
-    console.log(" -----> removeConversationHandler: ", conversationId);
+    console.log(' -----> removeConversationHandler: ', conversationId);
     const index = this.handlers.findIndex(i => i.conversationWith === conversationId);
     this.handlers.splice(index, 1);
   }
@@ -214,12 +214,12 @@ export class ChatManager {
    * @param conversationId
    */
   getConversationHandlerByConversationId(conversationId): any {
-    const resultArray = this.handlers.filter(function (handler) {
+    const resultArray = this.handlers.filter((handler) => {
       console.log('FILTRO::: ***', conversationId, handler.conversationWith);
-      return handler.conversationWith == conversationId;
+      return handler.conversationWith === conversationId;
     });
     console.log('getConversationHandlerByConversationId ***', conversationId, resultArray, this.handlers);
-    if (resultArray.length == 0) {
+    if (resultArray.length === 0) {
       return null;
     }
     return resultArray[0];
@@ -230,8 +230,8 @@ export class ChatManager {
    * dopo aver cancellato la reference per ogni handlers
    */
   setOffAllReferences() {
-    this.handlers.forEach(function (data) {
-      let item = data.ref;
+    this.handlers.forEach((data) => {
+      const item = data.ref;
       item.ref.off();
     });
     this.handlers = [];
@@ -243,15 +243,15 @@ export class ChatManager {
    * Salvo il CONVERSATIONS handler dopo averlo creato nella lista conversazioni
    */
   setConversationsHandler(handler) {
-    this.conversationsHandler = handler;
+    this.conversationsHandlerService = handler;
   }
 
   /**
    * elimino la reference dell'handler delle conversazioni
    */
   disposeConversationsHandler() {
-    console.log(" 2 - this.conversationsHandler:: ", this.conversationsHandler);
-    this.conversationsHandler.dispose();
+    console.log(' 2 - this.conversationsHandler:: ', this.conversationsHandlerService);
+    this.conversationsHandlerService.dispose();
   }
   /// END metodi gestione conversazioni ////
 
@@ -261,7 +261,7 @@ export class ChatManager {
    * inizio la sincronizzazione
    */
   initContactsSynchronizer() {
-    console.log(" initContactsSynchronizer:: ", this.contactsSynchronizer, this.tenant, this.loggedUser);
+    console.log(' initContactsSynchronizer:: ', this.contactsSynchronizer, this.tenant, this.loggedUser);
     if (!this.contactsSynchronizer) {
       this.contactsSynchronizer = this.chatContactsSynchronizer.initWithTenant(this.tenant, this.loggedUser);
       //this.contactsSynchronizer = this.createContactsSynchronizerForUser();